Refactor shared transposition logic and add editor preview controls
