+++ src/editor.scss
@@ -21,6 +21,13 @@
 		margin-top: 0.5em;
 	}
 
+	.chordpro-editor__transpose-value {
+		font-size: 0.85em;
+		color: #555;
+		min-width: 2.5rem;
+		text-align: center;
+	}
+
 	.chordpro-editor__preview-label {
 		font-size: 0.75em;
 		text-transform: uppercase;
